Added nodemap and armor override editors Added collider rotation scaling Added collider rotation setting to node editors Added TaskpoolOffload option (ini) Fixed misinterpreting collider types when loading JSON-serialized physics data Misc fixes and improvements

+++ CBP/CBP/Profiling.h
@@ -8,9 +8,9 @@ namespace CBP
         {
             long long avgTime;
             uint32_t avgActorCount;
-            long long avgStepRate;
-            uint32_t avgStepsPerUpdate;
-            float avgFrameTime;
+            double avgStepRate;
+            double avgStepsPerUpdate;
+            double avgFrameTime;
         };
 
     public:
@@ -37,7 +37,7 @@ namespace CBP
 
         uint32_t m_numActorsAccum;
         uint32_t m_numStepsAccum;
-        float m_frameTimeAccum;
+        double m_frameTimeAccum;
         uint32_t m_runCount;
 
         uint32_t m_uid;
